added dio as downloader (Android Only Yet)

+++ lib/providers/download_provider.dart
@@ -1,10 +1,6 @@
 import 'dart:io';
-import 'dart:isolate';
-import 'dart:ui';
 
 import 'package:flutter/material.dart';
-import 'package:flutter_downloader/flutter_downloader.dart';
-import 'package:hidayat/database/database.dart';
 import 'package:hidayat/models/bayan.dart';
 import 'package:hidayat/models/download_queue.dart';
 import 'package:hidayat/models/download_state.dart';
@@ -13,48 +9,6 @@ import 'package:path_provider/path_provider.dart';
 
 class DownloadProvider extends ChangeNotifier {
   final DownloadQueue _taskQueue = DownloadQueue();
-  ReceivePort _port = ReceivePort();
-
-  static void callback(String id, DownloadTaskStatus status, int progress) {
-    print("Downlaoding Callback");
-    print("id $id");
-    print("progress $progress");
-    print("status $status");
-    final SendPort send =
-        IsolateNameServer.lookupPortByName('downloader_send_port');
-    send.send([id, status, progress]);
-  }
-
-  DownloadProvider() {
-    IsolateNameServer.registerPortWithName(
-        _port.sendPort, 'downloader_send_port');
-    _port.listen((dynamic data) {
-      String id = data[0];
-      DownloadTaskStatus status = data[1];
-      int progress = data[2];
-      print("Downlaoding Callback in Provider");
-      print("id $id");
-      print("progress $progress");
-      print("status $status");
-      var downloadTaskState = _taskQueue.getDownloadStateByTaskId(id);
-      if (downloadTaskState == null) return;
-      downloadTaskState
-        ..progress = progress
-        ..status = status;
-      if (status == DownloadTaskStatus.complete) {
-        Bayan bayan = downloadTaskState.bayan;
-        downloadTaskState.bayan.filePath = downloadTaskState.path;
-
-        MySQLiteDatabase.getInstance().bayanDbHelper.updateBayan(bayan);
-        _taskQueue.removeTaskById(id);
-      } else if (status == DownloadTaskStatus.failed || progress == -1) {
-        String playlistId = _taskQueue.getDownloadStateByTaskId(id).playlistId;
-        _taskQueue.stopAllPlaylistTasks(playlistId);
-      }
-      notifyListeners();
-    });
-    FlutterDownloader.registerCallback(callback);
-  }
 
   ///returns `-1` if the playlist in not present.
   ///
@@ -64,7 +18,7 @@ class DownloadProvider extends ChangeNotifier {
   }
 
   double avgProgress(String playlistId) {
-    int avg = 0;
+    double avg = 0;
     List<DownloadTaskState> downloadStates =
         _taskQueue.getDownloadStateOfPlaylist(playlistId);
     if (downloadStates?.isEmpty ?? true) {
@@ -73,46 +27,34 @@ class DownloadProvider extends ChangeNotifier {
     downloadStates.forEach((element) {
       avg += element.progress;
     });
-    int length = downloadStates.length;
-    if (length < downloadStates.first.totalFiles) {
-      int diff = downloadStates.first.totalFiles - length;
-      avg += 100 * diff;
-    }
-    return avg / downloadStates.first.totalFiles;
+
+    return avg / downloadStates.length;
   }
 
   void downloadPlaylist(Playlist playlist) async {
     List<Bayan> bayans = playlist.bayans;
     Directory directory = await getExternalStorageDirectory();
-    // List<DownloadTaskState> tasks =
-    //     _taskQueue.getDownloadStateOfPlaylist(playlist.id);
-    // if (tasks == null) {
-    //   tasks = [];
-    //   _downloadQueue.putIfAbsent(playlist.id, () => tasks);
-    // }
+    String path = directory.path;
+    int pathSeparator =
+        Platform.pathSeparator.codeUnitAt(Platform.pathSeparator.length - 1);
+    if (path.codeUnitAt(path.length - 1) != pathSeparator) {
+      path = path + Platform.pathSeparator;
+    }
     for (int i = 0; i < bayans.length; i++) {
-      String path = directory.path;
-      int pathSeparator =
-          Platform.pathSeparator.codeUnitAt(Platform.pathSeparator.length - 1);
-      if (path.codeUnitAt(path.length - 1) != pathSeparator) {
-        path = path + Platform.pathSeparator;
-      }
       try {
         File file = new File(path + bayans[i].getUniqueFileName());
         if (await file.exists()) {
-          await file.delete();
+          if (bayans[i].filePath == file.path) {
+            continue;
+          }
+          else {
+            await file.delete();
+          }
         }
       } catch (ex) {
         print(ex);
       }
-
-      final taskId = await FlutterDownloader.enqueue(
-        url: bayans[i].link,
-        fileName: bayans[i].getUniqueFileName(),
-        showNotification: false,
-        requiresStorageNotLow: true,
-        savedDir: directory.path,
-      );
+      String taskId = "${DateTime.now().millisecondsSinceEpoch}";
       var downloadState = DownloadTaskState()
         ..bayan = bayans[i]
         ..progress = 0
@@ -122,7 +64,21 @@ class DownloadProvider extends ChangeNotifier {
         ..totalFiles = bayans.length
         ..playlistId = playlist.id;
       _taskQueue.addTask(downloadState);
+      // final taskId = await FlutterDownloader.enqueue(
+      //   url: bayans[i].link,
+      //   fileName: bayans[i].getUniqueFileName(),
+      //   showNotification: false,
+      //   requiresStorageNotLow: true,
+      //   savedDir: directory.path,
+      // );
     }
+    if (_taskQueue.getDownloadStateOfPlaylist(playlist.id)?.isEmpty ?? true) {
+      notifyListeners();
+      return;
+    }
+    _taskQueue.startDownloadQueue(playlist.id, () {
+      notifyListeners();
+    });
   }
 
   void removeAlltasksForPlaylist(String playlistId) {
